+++ src/Microsoft.Azure.Functions.Worker.Extensions.Mcp/Constants.cs
@@ -8,4 +8,10 @@ internal static class Constants
     public const string ToolInvocationContextKey = "ToolInvocationContext";
 
     public const string McpToolTriggerBindingType = "mcpToolTrigger";
+
+    public const string McpToolPropertyBindingType = "mcpToolProperty";
+
+    public const string McpToolPropertyType = "propertyType";
+
+    public const string McpToolPropertyName = "propertyName";
 }