+++ src/Microsoft.Azure.Functions.Worker.Extensions.Mcp/Converters/McpInputConversionHelper.cs
@@ -1,6 +1,8 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
+using System.Collections;
+using System.Collections.Concurrent;
 using System.ComponentModel;
 using System.Globalization;
 
@@ -8,6 +10,9 @@ namespace Microsoft.Azure.Functions.Worker.Extensions.Mcp.Converters;
 
 internal static class McpInputConversionHelper
 {
+    private static readonly ConcurrentDictionary<Type, Func<int, IList>> _listFactoryCache = new();
+    private static readonly ConcurrentDictionary<Type, Type> _elementTypeCache = new();
+
     public static bool TryConvertArgumentToTargetType(object? value, Type targetType, out object? result)
     {
         ArgumentNullException.ThrowIfNull(targetType);
@@ -32,25 +37,44 @@ internal static class McpInputConversionHelper
             return !targetType.IsValueType || Nullable.GetUnderlyingType(targetType) is not null;
         }
 
+        try
+        {
+            result = ConvertArgumentToTargetType(value, targetType);
+        }
+        catch { }
+
+        return result is not null || !targetType.IsValueType;
+    }
+
+    private static object? ConvertArgumentToTargetType(object? value, Type targetType)
+    {
+        if (value is null)
+        {
+            return null;
+        }
+
         targetType = Nullable.GetUnderlyingType(targetType) ?? targetType;
 
         if (targetType.IsEnum)
         {
             if (value is int || value is long)
             {
-                result = Enum.ToObject(targetType, value);
-                return true;
+                return Enum.ToObject(targetType, value);
             }
 
-            return Enum.TryParse(targetType, value.ToString(), ignoreCase: true, out result);
+            return Enum.Parse(targetType, value.ToString()!, ignoreCase: true);
+        }
+
+        if (IsSupportedCollectionType(targetType) && value is List<object?> inputCollection)
+        {
+            return ConvertToCollection(inputCollection, targetType);
         }
 
         if (value is IConvertible)
         {
             try
             {
-                result = Convert.ChangeType(value, targetType, CultureInfo.InvariantCulture);
-                return true;
+                return Convert.ChangeType(value, targetType, CultureInfo.InvariantCulture);
             }
             catch (Exception exc) when (exc is InvalidCastException || exc is FormatException) { }
         }
@@ -58,11 +82,131 @@ internal static class McpInputConversionHelper
         var typeConverter = TypeDescriptor.GetConverter(targetType);
         if (typeConverter is not null && typeConverter.CanConvertFrom(value.GetType()))
         {
-            result = typeConverter.ConvertFrom(null, CultureInfo.InvariantCulture, value);
-            return true;
+            return typeConverter.ConvertFrom(null, CultureInfo.InvariantCulture, value);
         }
 
-        result = value;
-        return result is not null || !targetType.IsValueType;
+        return null;
+    }
+
+    public static bool IsSupportedCollectionType(Type targetType)
+    {
+        return targetType != typeof(string)
+            && (targetType.IsArray || targetType.IsAssignableTo(typeof(IEnumerable)));
+    }
+
+    public static object? ConvertToCollection(List<object?> source, Type targetType)
+    {
+        ArgumentNullException.ThrowIfNull(targetType);
+
+        if (source is null)
+        {
+            return null;
+        }
+
+        var elementType = GetElementType(targetType)
+            ?? throw new InvalidOperationException($"Could not determine element type for {targetType}");
+
+        if (targetType.IsArray)
+        {
+            var array = Array.CreateInstance(elementType, source.Count);
+            for (int i = 0; i < source.Count; i++)
+            {
+                var item = ConvertElementIfNeeded(source[i], elementType);
+                array.SetValue(item, i);
+            }
+
+            return array;
+        }
+
+        var list = CreateTypedList(elementType, source.Count)!;
+        var listType = list.GetType();
+
+        foreach (var item in source)
+        {
+            var convertedItem = ConvertElementIfNeeded(item, elementType);
+            list.Add(convertedItem);
+        }
+
+        if (targetType.IsAssignableFrom(listType))
+        {
+            return list;
+        }
+
+        // Attempt to convert to target type via constructor that accepts IEnumerable<T>
+        var constructor = targetType.GetConstructor([typeof(IEnumerable<>).MakeGenericType(elementType)]);
+        if (constructor != null)
+        {
+            return constructor.Invoke([list]);
+        }
+
+        // Fall back to List<T>...
+        return list;
+    }
+
+    private static IList CreateTypedList(Type elementType, int capacity)
+    {
+        var factory = _listFactoryCache.GetOrAdd(
+            elementType,
+            static t =>
+            {
+                var constructedListType = typeof(List<>).MakeGenericType(t);
+                var ctor = constructedListType.GetConstructor(new[] { typeof(int) })!;
+                return (int cap) => (IList)ctor.Invoke(new object[] { cap });
+            });
+
+        return factory(capacity);
+    }
+
+    private static object? ConvertElementIfNeeded(object? element, Type targetType)
+    {
+        if (element is null)
+        {
+            return null;
+        }
+
+        if (targetType.IsAssignableFrom(element.GetType()))
+        {
+            return element;
+        }
+
+        if (TryConvertArgumentToTargetType(element, targetType, out var converted))
+        {
+            return converted;
+        }
+
+        return element;
+    }
+
+    private static Type? GetElementType(Type collectionType)
+    {
+        return _elementTypeCache.GetOrAdd(collectionType, static t => GetElementTypeCore(t));
+
+        static Type GetElementTypeCore(Type t)
+        {
+            if (t.IsArray)
+            {
+                return t.GetElementType()!;
+            }
+
+            if (MatchGenericEnumerable(t, null))
+            {
+                return t.GetGenericArguments()[0];
+            }
+
+            var enumerableType = t
+                .FindInterfaces(MatchGenericEnumerable, null)
+                .FirstOrDefault();
+
+            if (enumerableType is not null)
+            {
+                return enumerableType.GetGenericArguments()[0];
+            }
+
+            return typeof(object);
+        }
+
+        static bool MatchGenericEnumerable(Type type, object? _)
+            => type.IsGenericType && type.GetGenericTypeDefinition() == typeof(IEnumerable<>);
     }
 }
+